fix: test directory refactored
